Migrate Node binding to napi-rs
