+++ Cargo.toml
@@ -66,26 +66,32 @@ nodejs = ["console_error_panic_hook"]
 [dependencies]
 gluesql-core.workspace = true
 gluesql_memory_storage.workspace = true
-gluesql-web-storage.workspace = true
-gluesql-idb-storage.workspace = true
-gluesql-composite-storage.workspace = true
-
-wasm-bindgen = { version = "0.2.100" }
-wasm-bindgen-futures = "0.4.29"
-js-sys = "0.3"
 
 serde = "1"
 serde_json = "1"
-gloo-utils = { version = "0.1.6", features = ["serde"] }
 
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+gluesql-web-storage.workspace = true
+gluesql-idb-storage.workspace = true
+gluesql-composite-storage.workspace = true
+getrandom.workspace = true
+wasm-bindgen = { version = "0.2.100" }
+wasm-bindgen-futures = "0.4.29"
+js-sys = "0.3"
+gloo-utils = { version = "0.1.6", features = ["serde"] }
 # The `console_error_panic_hook` crate provides better debugging of panics by
 # logging them with `console.error`. This is great for development, but requires
 # all the `std::fmt` and `std::panicking` infrastructure, so isn't great for
 # code size when deploying.
 console_error_panic_hook = { version = "0.1.6", optional = true }
 
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-getrandom.workspace = true
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+futures-executor = "0.3"
+napi = { version = "3", default-features = false, features = ["napi4"] }
+napi-derive = "3"
+
+[build-dependencies]
+napi-build = "2"
 
 [dev-dependencies]
 test-suite.workspace = true
