+++ build.rs
@@ -1,0 +1,5 @@
+fn main() {
+    if std::env::var("CARGO_CFG_TARGET_ARCH").as_deref() != Ok("wasm32") {
+        napi_build::setup();
+    }
+}
